Simplify API to return price as BigDecimal
Clients don't need Money value object complexity - returning amount directly.

+++ src/test/groovy/com/urgoringo/mealkit/RecipesPricingSpec.groovy
@@ -17,6 +17,6 @@ class RecipesPricingSpec extends ApplicationSpecification {
 
         then: "system returns recipe pricing details"
             recipe.pricingCategory() == MEDIUM
-            recipe.price().toDouble() == 10.5
+            recipe.price() == 10.50
     }
 }
